Added ServiceResult.cs, project refactored to use it

+++ PriceObserver.Telegram/Concrete/Commands/Remove/RemoveCommandService.cs
@@ -1,6 +1,6 @@
-﻿using System;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using PriceObserver.Data.Repositories.Abstract;
+using PriceObserver.Model.Telegram.Commands;
 using PriceObserver.Telegram.Abstract.Commands.Remove;
 using Telegram.Bot.Types;
 using User = PriceObserver.Model.Data.User;
@@ -20,19 +20,25 @@ namespace PriceObserver.Telegram.Concrete.Commands.Remove
             _itemRepository = itemRepository;
         }
 
-        public async Task Process(Update update, User user)
+        public async Task<CommandExecutionResult> Process(Update update, User user)
         {
-            var parameters = _removeCommandParametersBuilder.Build(update);
-            var itemId = parameters.Id;
+            var parametersBuildResult = _removeCommandParametersBuilder.Build(update);
+
+            if (!parametersBuildResult.IsSuccess)
+                return CommandExecutionResult.Fail(parametersBuildResult.Error);
+            
+            var itemId = parametersBuildResult.Result.Id;
 
             var item = await _itemRepository.GetById(itemId);
             if (item == null)
-                throw new Exception("Item does not exist");
+                return CommandExecutionResult.Fail("Item does not exist");
             
             if (item.UserId != user.Id)
-                throw new Exception("This item is not yours");
+                return CommandExecutionResult.Fail("This item is not yours");
 
             await _itemRepository.Delete(item);
+
+            return CommandExecutionResult.Success("Item deleted successfully!");
         }
     }
 }
